인앱 업데이트 로직 구현

+++ repository/src/main/java/com/skfo763/repository/test/TestCheckPointGenerator.kt
@@ -9,8 +9,8 @@ class TestCheckPointGenerator {
     fun getRandomCheckPoint(): CheckPoint {
         return CheckPoint(
             checkPointIdx = Random.nextInt(0, 100000),
-            latitude = Random.nextDouble(37.512999, 37.715133).toFloat(),
-            longitude = Random.nextDouble(127.102268, 127.269311).toFloat(),
+            latitude = Random.nextDouble(37.512999, 37.715133),
+            longitude = Random.nextDouble(127.102268, 127.269311),
             address = CheckInAddress(
                 "대한민국",
                 "서울특별시",
